Renaming the components for the routing to work.
