+++ StudentResourceHub/src/app/about/about.component.ts
@@ -1,11 +1,15 @@
-import { Component } from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 @Component({
-  selector: 'app-about',
-  imports: [],
-  templateUrl: './about.component.html',
-  styleUrl: './about.component.css'
+  selector: 'app-about',  // Updated selector
+  standalone: true,
+  templateUrl: './about.component.html',  // Updated template URL
+  styleUrls: ['./about.component.css']  // Updated styles URL
 })
 export class AboutComponent {
+  @Input() resource: { title: string; description: string; category: string; imageUrl: string } | null = null;
 
+  goBack(): void {
+    console.log('Navigating back to resource list');
+  }
 }
