--- StudentResourceHub/src/app/add-resource-form/add-resource-form.component.html
@@ -1,48 +1,0 @@
-<div class="card p-3">
-    <h3>Add New Resource</h3>
-    <form (ngSubmit)="addResource()" #resourceForm="ngForm">
-      <div class="mb-3">
-        <label for="title" class="form-label">Title</label>
-        <input 
-          type="text" 
-          id="title" 
-          class="form-control" 
-          [(ngModel)]="resource.title" 
-          name="title" 
-          required 
-          [ngClass]="{ 'is-invalid': resourceForm.controls['title']?.invalid && resourceForm.controls['title']?.touched }">
-      </div>
-  
-      <div class="mb-3">
-        <label for="description" class="form-label">Description</label>
-        <textarea 
-          id="description" 
-          class="form-control" 
-          [(ngModel)]="resource.description" 
-          name="description" 
-          required
-          [ngClass]="{ 'is-invalid': resourceForm.controls['description']?.invalid && resourceForm.controls['description']?.touched }">
-        </textarea>
-      </div>
-  
-      <div class="mb-3">
-        <label for="category" class="form-label">Category</label>
-        <input 
-          type="text" 
-          id="category" 
-          class="form-control" 
-          [(ngModel)]="resource.category" 
-          name="category" 
-          required
-          [ngClass]="{ 'is-invalid': resourceForm.controls['category']?.invalid && resourceForm.controls['category']?.touched }">
-      </div>
-  
-      <button 
-        type="submit" 
-        class="btn btn-primary" 
-        [disabled]="resourceForm.invalid">
-        Add Resource
-      </button>
-    </form>
-  </div>
-
