--- StudentResourceHub/src/app/add-resource-form/add-resource-form.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AddResourceFormComponent } from './add-resource-form.component';
-
-describe('AddResourceFormComponent', () => {
-  let component: AddResourceFormComponent;
-  let fixture: ComponentFixture<AddResourceFormComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [AddResourceFormComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(AddResourceFormComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
