--- StudentResourceHub/src/app/categories/design/design.component.html
@@ -1,1 +1,0 @@
-<p>design works!</p>
