--- StudentResourceHub/src/app/categories/design/design.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-design',
-  imports: [],
-  templateUrl: './design.component.html',
-  styleUrl: './design.component.css'
-})
-export class DesignComponent {
-
-}
