--- StudentResourceHub/src/app/categories/math/math.component.html
@@ -1,1 +1,0 @@
-<p>math works!</p>
