--- StudentResourceHub/src/app/categories/math/math.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { MathComponent } from './math.component';
-
-describe('MathComponent', () => {
-  let component: MathComponent;
-  let fixture: ComponentFixture<MathComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [MathComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(MathComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
