--- StudentResourceHub/src/app/categories/math/math.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-math',
-  imports: [],
-  templateUrl: './math.component.html',
-  styleUrl: './math.component.css'
-})
-export class MathComponent {
-
-}
