--- StudentResourceHub/src/app/categories/programming/programming.component.html
@@ -1,1 +1,0 @@
-<p>programming works!</p>
