--- StudentResourceHub/src/app/categories/programming/programming.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ProgrammingComponent } from './programming.component';
-
-describe('ProgrammingComponent', () => {
-  let component: ProgrammingComponent;
-  let fixture: ComponentFixture<ProgrammingComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ProgrammingComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(ProgrammingComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
