--- StudentResourceHub/src/app/categories/programming/programming.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-programming',
-  imports: [],
-  templateUrl: './programming.component.html',
-  styleUrl: './programming.component.css'
-})
-export class ProgrammingComponent {
-
-}
