+++ StudentResourceHub/src/app/home/home.component.css
@@ -1,0 +1,8 @@
+.card {
+    transition: transform 0.2s;
+  }
+  
+  .card:hover {
+    transform: scale(1.03);
+  }
+
