--- StudentResourceHub/src/app/navigation-menu/navigation-menu.component.html
@@ -1,1 +1,0 @@
-<p>navigation-menu works!</p>
