--- StudentResourceHub/src/app/resource-category/resource-category.component.ts
@@ -1,44 +1,0 @@
-import { Component, Input, OnInit } from '@angular/core';
-
-interface Resource {
-  id: number;
-  title: string;
-  description: string;
-  category: string;
-}
-
-interface Category {
-  title: string;
-  description: string;
-}
-
-@Component({
-  selector: 'app-resource-category',
-  templateUrl: './resource-category.component.html',
-  styleUrls: ['./resource-category.component.css']
-})
-export class ResourceCategoryComponent implements OnInit {
-  @Input() category: Category | null = null; // Category passed as input
-  resources: Resource[] = [];
-
-  allResources: Resource[] = [
-    { id: 1, title: 'Angular Basics', description: 'Introduction to Angular', category: 'Programming' },
-    { id: 2, title: 'UI/UX Design Principles', description: 'Understanding UI/UX', category: 'Design' },
-    { id: 3, title: 'Calculus Guide', description: 'Math resources for beginners', category: 'Math' },
-    { id: 4, title: 'React Introduction', description: 'Learn React', category: 'Programming' }
-  ];
-
-  ngOnInit(): void {
-    if (this.category) {
-      this.filterResourcesByCategory(this.category.title);
-    }
-  }
-
-  filterResourcesByCategory(category: string): void {
-    this.resources = this.allResources.filter(resource => resource.category === category);
-  }
-
-  viewDetails(id: number): void {
-    console.log(`Viewing details for resource ${id}`);
-  }
-}
