--- StudentResourceHub/src/app/resource-list/resource-list.component.css
@@ -1,8 +1,0 @@
-.card {
-    transition: transform 0.2s;
-  }
-  
-  .card:hover {
-    transform: scale(1.03);
-  }
-
